refactor: rename safe_selector to safe_navigation, update imports
