+++ apps/web/cypress/e2e/pages/safe_navigation.pages.js
